ncu-ci: parse console text of the first failed phase (#351)
Previously we try to detect the type of the failed phase
and return a generic build failure when the job failed in a
non-test phase. This is not really necessary - the console
text parser is capable of detecting the actual build failure
from even the console text of the build phase. Also hard-coding
the test phase name makes the detection fragile when there
are configuration changes in the Jenkins. So we should just
skip the detection.