fix: use case-insensitive string comparison for fast-track approvals (#658)
I noticed today that a PR that had sufficient fast-track approvals was
being rejected as needing one more approval. This is because one of the
approvers' had a GitHub handle where the case in the README
(`linkgoron`) did not match the case returned by GitHub (`Linkgoron`).
Since GitHub handles are case-insensitive, this change makes the
comparison case-insensitive.

One might be tempted to think that we need to do the same for the list
of TSC members, but handles are never compared there so it is not
necessary.